refactor(admin): consolidate shared admin components module
Reusable pieces such as the connection credential input, tool-call badge, workspace task list and the service/model config form lived in view/, composables/ and styles/, while app-shell logic sat in composables/. Fold them into admin/components with their scss beside each component so shared UI evolves as one module.
