+++ webui/src/admin/components/ConnectionCredentialInput.vue
