+++ webui/src/admin/components/admin-app.scss
