+++ webui/src/admin/components/service-model-config.scss
