+++ webui/src/admin/components/tool-call-badge.scss
