+++ webui/src/admin/components/useAdminClipboard.ts
