+++ webui/src/admin/components/useToolCallBadge.ts
